+++ package.json
@@ -71,6 +71,7 @@
     "@fortawesome/free-solid-svg-icons": "^5.11.2",
     "@fortawesome/react-fontawesome": "^0.1.4",
     "@percy/storybook": "^3.2.0",
+    "@plotly/dash-component-plugins": "^1.0.1",
     "@plotly/webpack-dash-dynamic-import": "^1.0.0",
     "@storybook/cli": "^5.1.11",
     "@storybook/react": "^5.1.11",
