+++ src/dash-table/dash/DataTable.js
@@ -1,14 +1,16 @@
 import * as R from 'ramda';
-import React, { Component } from 'react';
+import React, { Component, lazy } from 'react';
 import PropTypes from 'prop-types';
-
-import RealTable from 'dash-table/components/Table';
+import { asyncDecorator } from '@plotly/dash-component-plugins';
 
 import Logger from 'core/Logger';
 
 import genRandomId from 'dash-table/utils/generate';
 import isValidProps from './validate';
 import Sanitizer from './Sanitizer';
+import LazyLoader from 'dash-table/LazyLoader';
+
+const RealDataTable = asyncDecorator(DataTable, LazyLoader.table);
 
 /**
  * Dash DataTable is an interactive table component designed for
@@ -20,22 +22,12 @@ import Sanitizer from './Sanitizer';
  * is completely customizable through its properties.
  */
 export default class DataTable extends Component {
-    constructor(props) {
-        super(props);
-        let id;
-        this.getId = () => (id = id || genRandomId('table-'));
-        this.sanitizer = new Sanitizer();
-    }
-
     render() {
-        if (!isValidProps(this.props)) {
-            return (<div>Invalid props combination</div>);
-        }
-
-        const sanitizedProps = this.sanitizer.sanitize(this.props);
-        return this.props.id ?
-            (<RealTable {...sanitizedProps} />) :
-            (<RealTable {...sanitizedProps} id={this.getId()} />);
+        return (
+            <Suspense fallback={null}>
+                <RealDataTable {...this.props} />
+            </Suspense>
+        );
     }
 }
